+++ src/app/(auth)/signup/page.js
@@ -1,5 +1,5 @@
 import Heading from "@/components/Heading";
-import SignUpForm from "@/components/SignUpForm";
+import SignUpForm from "@/components/SignUpForm/SignUpForm";
 
 const SignUp = () => (
   <>
