+++ src/components/SignUpForm/SignUpFormStepOne.jsx
@@ -1,0 +1,25 @@
+"use client";
+
+import "react-toastify/dist/ReactToastify.css";
+
+import AuthInput from "@/components/AuthInput";
+
+const SignUpFormStepOne = ({ fields, errors, control }) => {
+  return (
+    <div className="space-y-6">
+      {fields.map(({ children, name, ...otherProps }) => (
+        <AuthInput
+          key={name}
+          name={name}
+          control={control}
+          errors={errors}
+          {...otherProps}
+        >
+          {children}
+        </AuthInput>
+      ))}
+    </div>
+  );
+};
+
+export default SignUpFormStepOne;
